backend/ai-services/requirements.txt
# Core dependencies
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# AI and ML
google-generativeai==0.3.2
langchain==0.1.0
openai==1.6.1
tiktoken==0.5.2

# Database
asyncpg==0.29.0
pgvector==0.2.4
aioredis==2.0.1
sqlalchemy==2.0.25
alembic==1.13.1

# Web scraping and processing
beautifulsoup4==4.12.2
scrapy==2.11.0
html2text==2020.1.16
markdown==3.5.1

# Async and concurrency
aiohttp==3.9.1
asyncio==3.4.3

# Data processing
numpy==1.26.3
pandas==2.1.4
pydantic==2.5.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Logging and monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
ipython==8.19.0